Rename quantize operator to quantize_bf16_i8 for naming consistency
Align operator directory name with the kernel file convention
(quantize_bf16_i8.cc). This establishes a consistent full
input_output type naming pattern for quantization operators.
